Simplify build structure
