--- okhttp-futureapi/README.md
@@ -1,4 +1,0 @@
-OkHttp Future API
-=================
-
-Future API bindings e.g. Android Q and Java 9.
