--- okhttp-futureapi/build.gradle
@@ -1,24 +1,0 @@
-apply plugin: 'com.vanniktech.maven.publish'
-
-jar {
-  manifest {
-    attributes('Automatic-Module-Name': 'okhttp3.futureapi')
-  }
-}
-
-dependencies {
-  api deps.kotlinStdlib
-  compileOnly deps.jsr305
-  compileOnly deps.android
-  compileOnly deps.conscrypt
-
-  testImplementation deps.junit
-  testImplementation deps.assertj
-}
-
-afterEvaluate { project ->
-  project.tasks.dokka {
-    outputDirectory = "$rootDir/docs/4.x"
-    outputFormat = 'gfm'
-  }
-}
