--- okhttp-futureapi/gradle.properties
@@ -1,3 +1,0 @@
-POM_ARTIFACT_ID=okhttp-futureapi
-POM_NAME=okhttp-futureapi
-POM_PACKAGING=jar
