--- okhttp-futureapi/src/main/java/okhttp3/internal/package-info.kt
@@ -1,2 +1,0 @@
-/** Bindings for Future API for OkHttp. */
-package okhttp3.internal.futureapi
