+++ okhttp/build.gradle
@@ -25,6 +25,7 @@ task copyJavaTemplates(type: Copy) {
 dependencies {
   api deps.okio
   api deps.kotlinStdlib
+  compileOnly deps.android
   compileOnly deps.conscrypt
   compileOnly deps.openjsse
   compileOnly deps.jsr305
